requirements.txt
asgiref==3.8.1
awsebcli==3.21.0
blessed==1.20.0
blinker==1.8.2
botocore==1.35.40
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.1
pathspec==0.10.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
six==1.16.0
SQLAlchemy==2.0.35
sqlparse==0.5.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.0.4
